implement adding comments

+++ app/controllers/articles_controller.rb
@@ -22,6 +22,8 @@ class ArticlesController < ApplicationController
   end
   
   def show
+    @comment = @article.comments.build
+    @comments = @article.comments
   end
   
   def edit
@@ -47,9 +49,14 @@ class ArticlesController < ApplicationController
   end
   
   def destroy
-    if @article.destroy
-      flash[:success] = "Article has been deleted"
-      redirect_to articles_path
+    unless @article.user == current_user
+      flash[:danger] = "You can only delete your own article"
+      redirect_to root_path
+    else
+      if @article.destroy
+        flash[:success] = "Article has been deleted"
+        redirect_to articles_path
+      end
     end
   end
